requirements.txt
absl-py==0.7.1
aniso8601==8.0.0
ansible==2.8.1
ansible-lint==4.1.0
appdirs==1.4.3
appnope==0.1.0
asn1crypto==0.24.0
astor==0.8.0
astroid==2.2.5
attrs==19.1.0
Automat==0.7.0
backcall==0.1.0
beautifulsoup4==4.7.1
black==19.3b0
bleach==3.1.0
bokeh==1.2.0
captcha-solver==0.1.5
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.7
cssselect==1.1.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.1
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
gast==0.2.2
get==2019.4.13
grpcio==1.21.1
gunicorn==19.9.0
h5py==2.9.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.20
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lazy-object-proxy==1.4.1
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mock==3.0.5
mpmath==1.1.0
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.16.2
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parsel==1.5.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
post==2019.4.13
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.8.0
psycopg2==2.8.3
ptyprocess==0.6.0
public==2019.4.13
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
PyDispatcher==2.0.5
Pygments==2.4.2
PyHamcrest==1.9.0
pylint==2.3.1
PyMySQL==0.9.3
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyrsistent==0.15.4
pytesseract==0.3.0
python-dateutil==2.8.0
python-docx==0.8.10
python-dotenv==0.10.2
pytz==2019.1
PyYAML==5.1
pyzmq==18.1.0
qtconsole==4.5.5
query-string==2019.4.13
queuelib==1.5.0
request==2019.4.13
requests==2.21.0
rope==0.14.0
ruamel.yaml==0.15.97
scikit-learn==0.21.2
scipy==1.3.0
Scrapy==1.7.3
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.12.0
sklearn==0.0
snownlp==0.12.3
soupsieve==1.9.1
SQLAlchemy==1.3.8
sympy==1.4
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
toml==0.10.0
tornado==6.0.2
traitlets==4.3.2
Twisted==19.7.0
typed-ast==1.3.5
urllib3==1.24.1
w3lib==1.21.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.5.1
wrapt==1.11.1
zope.interface==4.6.0
